Converted "int" to "short" for "RelationOrder" in "RelationSet_Relation".

+++ src/gr/csri/poeticon/praxicon/db/dao/implSQL/RelationDaoImpl.java
@@ -72,7 +72,7 @@ public class RelationDaoImpl extends JpaDao<Long, Relation> implements
                 r.setType(tmpType);
             }
             RelationSet rs = new RelationSet();
-            rs.addRelation(r, 0);
+            rs.addRelation(r, (short) 0);
         }
         return res;
     }
